refactor: 공통 컴포저블 분리 및 미사용 컴포넌트 제거
- `core:designsystem` 모듈에서 더 이상 사용되지 않는 `BottariBottomSheet` 컴포저블을 삭제했습니다.
- `presentation` 모듈에서만 사용되던 `BottariRenameTextField`를 제거했습니다.
- `core:ui` 모듈에서 사용하지 않는 `BottariSearchBar`를 삭제했습니다.
- `BottariHashChipSearchBar`의 패키지 경로를 `component`로 이동하고, 관련 참조를 업데이트했습니다.
- `HomeTopAppBar`가 `designsystem`의 공통 `BottariTopAppBar`를 재사용하도록 리팩토링했습니다.
- `BottariTopAppBar`의 예시 코드에서 `ArrowBack` 아이콘을 `AutoMirrored` 버전으로 변경했습니다.
